tock-register-interface: document & move bitfield types + macros
