+++ libraries/riscv-csr/src/csr.rs
@@ -2,8 +2,9 @@
 
 use core::marker::PhantomData;
 
+use tock_registers::fields::Field;
 use tock_registers::interfaces::{Readable, Writeable};
-use tock_registers::registers::{Field, IntLike, RegisterLongName};
+use tock_registers::registers::{IntLike, RegisterLongName};
 
 pub const MINSTRETH: usize = 0xB82;
 pub const MINSTRET: usize = 0xB02;
