+++ libraries/tock-register-interface/src/lib.rs
@@ -5,6 +5,7 @@
 #![feature(const_fn_trait_bound)]
 #![no_std]
 
+pub mod fields;
 pub mod interfaces;
 pub mod macros;
 pub mod registers;
