+++ libraries/tock-register-interface/src/registers.rs
@@ -53,17 +53,12 @@
 //! ------
 //! - Shane Leonard <shanel@stanford.edu>
 
-// The register interface uses `+` in a way that is fine for bitfields, but
-// looks unusual (and perhaps problematic) to a linter. We just ignore those
-// lints for this file.
-#![allow(clippy::suspicious_op_assign_impl)]
-#![allow(clippy::suspicious_arithmetic_impl)]
-
 use core::cell::UnsafeCell;
 use core::fmt;
 use core::marker::PhantomData;
-use core::ops::{Add, AddAssign, BitAnd, BitOr, BitOrAssign, Not, Shl, Shr};
+use core::ops::{BitAnd, BitOr, BitOrAssign, Not, Shl, Shr};
 
+use crate::fields::{Field, FieldValue};
 use crate::interfaces::{Readable, Writeable};
 
 /// IntLike properties needed to read/write/modify a register.
@@ -384,183 +379,6 @@ From_impl_for!(u32);
 From_impl_for!(u64);
 From_impl_for!(u128);
 From_impl_for!(usize);
-
-/// Specific section of a register.
-///
-/// For the Field, the mask is unshifted, ie. the LSB should always be set.
-pub struct Field<T: IntLike, R: RegisterLongName> {
-    pub mask: T,
-    pub shift: usize,
-    associated_register: PhantomData<R>,
-}
-
-impl<T: IntLike, R: RegisterLongName> Field<T, R> {
-    pub const fn new(mask: T, shift: usize) -> Field<T, R> {
-        Field {
-            mask: mask,
-            shift: shift,
-            associated_register: PhantomData,
-        }
-    }
-
-    #[inline]
-    pub fn read(self, val: T) -> T {
-        (val & (self.mask << self.shift)) >> self.shift
-    }
-
-    #[inline]
-    /// Check if one or more bits in a field are set
-    pub fn is_set(self, val: T) -> bool {
-        val & (self.mask << self.shift) != T::zero()
-    }
-
-    #[inline]
-    /// Read value of the field as an enum member
-    pub fn read_as_enum<E: TryFromValue<T, EnumType = E>>(self, val: T) -> Option<E> {
-        E::try_from(self.read(val))
-    }
-}
-
-// #[derive(Copy, Clone)] won't work here because it will use
-// incorrect bounds, as a result of using a PhantomData over the
-// generic R. The PhantomData<R> implements Copy regardless of whether
-// R does, but the #[derive(Copy, Clone)] generates
-//
-//    #[automatically_derived]
-//    #[allow(unused_qualifications)]
-//    impl<T: ::core::marker::Copy + IntLike,
-//         R: ::core::marker::Copy + RegisterLongName>
-//            ::core::marker::Copy for Field<T, R> {}
-//
-// , so Field will only implement Copy if R: Copy.
-//
-// Manually implementing Clone and Copy works around this issue.
-//
-// Relevant Rust issue: https://github.com/rust-lang/rust/issues/26925
-impl<T: IntLike, R: RegisterLongName> Clone for Field<T, R> {
-    fn clone(&self) -> Self {
-        Field {
-            mask: self.mask,
-            shift: self.shift,
-            associated_register: self.associated_register,
-        }
-    }
-}
-impl<T: IntLike, R: RegisterLongName> Copy for Field<T, R> {}
-
-macro_rules! Field_impl_for {
-    ($type:ty) => {
-        impl<R: RegisterLongName> Field<$type, R> {
-            pub fn val(&self, value: $type) -> FieldValue<$type, R> {
-                FieldValue::<$type, R>::new(self.mask, self.shift, value)
-            }
-        }
-    };
-}
-
-Field_impl_for!(u8);
-Field_impl_for!(u16);
-Field_impl_for!(u32);
-Field_impl_for!(u64);
-Field_impl_for!(u128);
-Field_impl_for!(usize);
-
-/// Values for the specific register fields.
-///
-/// For the FieldValue, the masks and values are shifted into their actual
-/// location in the register.
-#[derive(Copy, Clone)]
-pub struct FieldValue<T: IntLike, R: RegisterLongName> {
-    mask: T,
-    pub value: T,
-    associated_register: PhantomData<R>,
-}
-
-macro_rules! FieldValue_impl_for {
-    ($type:ty) => {
-        // Necessary to split the implementation of new() out because the bitwise
-        // math isn't treated as const when the type is generic.
-        // Tracking issue: https://github.com/rust-lang/rfcs/pull/2632
-        impl<R: RegisterLongName> FieldValue<$type, R> {
-            pub const fn new(mask: $type, shift: usize, value: $type) -> Self {
-                FieldValue {
-                    mask: mask << shift,
-                    value: (value & mask) << shift,
-                    associated_register: PhantomData,
-                }
-            }
-        }
-
-        // Necessary to split the implementation of From<> out because of the orphan rule
-        // for foreign trait implementation (see [E0210](https://doc.rust-lang.org/error-index.html#E0210)).
-        impl<R: RegisterLongName> From<FieldValue<$type, R>> for $type {
-            fn from(val: FieldValue<$type, R>) -> $type {
-                val.value
-            }
-        }
-    };
-}
-
-FieldValue_impl_for!(u8);
-FieldValue_impl_for!(u16);
-FieldValue_impl_for!(u32);
-FieldValue_impl_for!(u64);
-FieldValue_impl_for!(u128);
-FieldValue_impl_for!(usize);
-
-impl<T: IntLike, R: RegisterLongName> FieldValue<T, R> {
-    /// Get the raw bitmask represented by this FieldValue.
-    #[inline]
-    pub fn mask(&self) -> T {
-        self.mask as T
-    }
-
-    #[inline]
-    pub fn read(&self, field: Field<T, R>) -> T {
-        field.read(self.value)
-    }
-
-    /// Modify fields in a register value
-    #[inline]
-    pub fn modify(self, val: T) -> T {
-        (val & !self.mask) | self.value
-    }
-
-    /// Check if any specified parts of a field match
-    #[inline]
-    pub fn matches_any(&self, val: T) -> bool {
-        val & self.mask != T::zero()
-    }
-
-    /// Check if all specified parts of a field match
-    #[inline]
-    pub fn matches_all(&self, val: T) -> bool {
-        val & self.mask == self.value
-    }
-}
-
-// Combine two fields with the addition operator
-impl<T: IntLike, R: RegisterLongName> Add for FieldValue<T, R> {
-    type Output = Self;
-
-    #[inline]
-    fn add(self, rhs: Self) -> Self {
-        FieldValue {
-            mask: self.mask | rhs.mask,
-            value: self.value | rhs.value,
-            associated_register: PhantomData,
-        }
-    }
-}
-
-// Combine two fields with the += operator
-impl<T: IntLike, R: RegisterLongName> AddAssign for FieldValue<T, R> {
-    #[inline]
-    fn add_assign(&mut self, rhs: FieldValue<T, R>) {
-        self.mask |= rhs.mask;
-        self.value |= rhs.value;
-    }
-}
 
 #[cfg(not(feature = "no_std_unit_tests"))]
 #[cfg(test)]
